[side-project] update personal information refactoring and id type convert and test commit

+++ Projects/SideProject/common/gc_storage_image.go
@@ -158,7 +158,7 @@ func SaveUserImgFile(args ArgsUpdateJoinUserInfo) (string, error) {
 }
 
 //DeleteUserImgFile is function to check thage the user image folder exists in storage.
-func DeleteUserImgFile(ctx context.Context, userid int, database *sql.DB) error {
+func DeleteUserImgFile(ctx context.Context, userid int64, database *sql.DB) error {
 
 	var userimglink string
 	stmt, err := database.Prepare(`SELECT image_link FROM user WHERE id = ?`)
